backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pandas
pymupdf
numpy